+++ admin/blogs.php
@@ -270,7 +270,7 @@ adminHeader('Správa blogů');
           <a href="blog.php?blog=<?= (int)$blog['id'] ?>" class="btn">Články blogu</a>
           <a href="blog_cats.php?blog_id=<?= (int)$blog['id'] ?>" class="btn">Kategorie blogu</a>
           <a href="blog_tags.php?blog_id=<?= (int)$blog['id'] ?>" class="btn">Štítky blogu</a>
-          <a href="blog_pages.php?blog_id=<?= (int)$blog['id'] ?>" class="btn">Stránky blogu</a>
+          <a href="blog_pages.php?blog_id=<?= (int)$blog['id'] ?>" class="btn">Stránky a odkazy blogu</a>
           <a href="blog_members.php?blog_id=<?= (int)$blog['id'] ?>" class="btn">Tým blogu</a>
           <form action="blog_blog_delete.php" method="post" class="blog-inline-form">
             <input type="hidden" name="csrf_token" value="<?= h(csrfToken()) ?>">